crypto: move encrypted to-device tests to new file
